Create jest config at install
Possibly this should be done at postinstall but since the base generator already defines a script there, I'm putting this here as a simple way to avoid overwriting that.
Also, we still need to run npm install --save-dev eslint-plugin-jest afterwards and add 'jest' to the plugins in the eslint config afterwards.
The reason this is necessary is because the eslint config init script doesn't present options for jest. We should probably use a subgenerator to deal with this.