Include dependencies for IDE integrated eslinting
This action changes the instructions required in the README.md.

The inclusion of these dependencies is a workaround to the fact that npm
seems to be misbehaving quite badly by removing dependencies which are
required by the project. This was proven during the manual install on
`eslint-plugin-jest` which would result in the removal of lots of other
eslint packages including the airbnb config base. I don't know whether a
similar thing is happening when the eslint config generator invites us
to install missing dependencies, but I'd rather not risk it in a process
which is intended to be used as a reliable base to new projects.